Add session check

+++ app/Http/Controllers/CategoryController.php
@@ -13,16 +13,23 @@ class CategoryController extends Controller
     public function showCategories(Request $request)
     {
         $fields = $request->validate(['userId'=>'']);
-        $category = Category::where(['userId'=>$fields['userId']])->get(['id','userId','categoryName']);
-        $categoryCount = DB::table('categories')
-            ->leftJoin('contact_to_categories','contact_to_categories.categoryId','=','categories.id')
-            ->leftJoin('contacts','contacts.id','=','contact_to_categories.contactId')
-            ->select('categories.categoryName', DB::raw('count(contacts.id)'))
-            ->where('categories.userId',$fields['userId'])
-            ->groupBy('categories.categoryName')
-            ->get();
-        $response = ['category'=>$category,'count'=>$categoryCount];
-        return response($response,201);
+        if($request->session()->exists('userId'))
+        {
+            $category = Category::where(['userId'=>$fields['userId']])->get(['id','userId','categoryName']);
+            $categoryCount = DB::table('categories')
+                ->leftJoin('contact_to_categories','contact_to_categories.categoryId','=','categories.id')
+                ->leftJoin('contacts','contacts.id','=','contact_to_categories.contactId')
+                ->select('categories.categoryName', DB::raw('count(contacts.id)'))
+                ->where('categories.userId',$fields['userId'])
+                ->groupBy('categories.categoryName')
+                ->get();
+            $response = ['category'=>$category,'count'=>$categoryCount];
+            return response($response,201);
+        }
+        else
+        {
+            return response()->view('404');
+        }
     }
 
     //Add new category
@@ -32,8 +39,15 @@ class CategoryController extends Controller
             'categoryName'=>'required',
             'userId'=>'required'
         ]);
-        Category::create(['userId' => $fields['userId'], 'categoryName' => $fields['categoryName']]);
-        return $this->showCategories($request);
+        if($request->session()->exists('userId'))
+        {
+            Category::create(['userId' => $fields['userId'], 'categoryName' => $fields['categoryName']]);
+            return $this->showCategories($request);
+        }
+        else
+        {
+            return response()->view('404');
+        }
     }
 
     //Update current category
@@ -43,18 +57,32 @@ class CategoryController extends Controller
             'id'=>'required',
             'categoryName'=>'required'
         ]);
-        Category::where(['id' => $fields['id']])->update([
-            'categoryName' => $fields['categoryName'],
+        if($request->session()->exists('userId'))
+        {
+            Category::where(['id' => $fields['id']])->update([
+                'categoryName' => $fields['categoryName'],
             ]);
-        return $this->showCategories($request);
+            return $this->showCategories($request);
+        }
+        else
+        {
+            return response()->view('404');
+        }
     }
 
     //Delete current category
     public function deleteCategory(Request $request)
     {
         $fields = $request->validate(['id'=>'required']);
-        Category::where(['id' => $fields['id']])->delete();
-        return $this->showCategories($request);
+        if($request->session()->exists('userId'))
+        {
+            Category::where(['id' => $fields['id']])->delete();
+            return $this->showCategories($request);
+        }
+        else
+        {
+            return response()->view('404');
+        }
     }
 
 }
